Fix the two last unit tests
There is an issue with decrypting mixed data, where the output buffer is
overwritten with previously decrypted data. On the case of miss-aligned
mixed data part of the clear portion of the outbuffer is lost. This fix adds a
temporary buffer to where the data that will be overwritten is saved. This
buffer is then used to restore the lost clear data.
Because we do not now the subsample composition, this performs redundant
memory copies on non-mixed miss-aligned data.